Add optional post commands to toolchains
This adds the following toolchain variables:
 - post_asm
 - post_compile
 - post_alink
 - post_solink
 - post_solink_module

These variables allow for commands to be executed after an assembly,
compile, or link command. In addition, `*_outputs` array variables are
added to specify additional outputs for each assembly, compile, or link
command. However, multiple assembly or compile outputs are not suported
by ninja (yet?).

The additional `*_outputs` array variables are as follows:
 - asm_outputs
 - compile_outputs
 - alink_outputs
 - solink_outputs
 - solink_module_outputs