Check for invalid ninja targets
This updates `bam-build` and `bam-clean` to check for valid ninja
targets before dispatching to `ninja` to perform a build or clean. If a
valid ninja target is not found, then an error is reported and the
command exits with an error.

Since target validation is shared across multiple commands, move target
resolution from a target name or label pattern into a common utility
function.