Refactor tests to be more robust
Update tests to be more robust against changes to `gn` output. For
example, `gn` outputs extra lines indicating the status of different
files it generates, such as JSON projects and compile_commands. As these
status lines are added or removed based on the command-line options sent
to `gn gen`, tests that are not directly testing the behavior of `gn
gen` should not need to be modified when this output changes.